+++ philo_bonus/fork.c
@@ -18,13 +18,13 @@ void	release_fork(sem_t *forks)
 	sem_post(forks);
 }
 
-void	release_forks(sem_t *forks, int num)
+void	release_forks(sem_t *forks, long num)
 {
 	while (num--)
 		release_fork(forks);
 }
 
-int	release_forks_and_rtn_err(sem_t *forks, int num)
+int	release_forks_and_rtn_err(sem_t *forks, long num)
 {
 	release_forks(forks, num);
 	return (-1);
